mysql_cli_agent/requirements.txt
mysql-connector-python
pytest
mcp[cli]
